[refactor, feat] JwtToken 구조 refactoring #9 RefreshToken Entity extends Auditing 클래스 JwtUtil 삭제
